Update primitive for textures

+++ includes/Primitive.h
@@ -94,7 +94,7 @@ public:
             {
                 number = std::to_string(emmisionNR++);
             }
-            shader.setFloat(("material." + name + number).c_str(), i);
+            shader.setInt(("material." + name + number).c_str(), i);
             glBindTexture(GL_TEXTURE_2D, textures[i].id);
         }
         glActiveTexture(GL_TEXTURE0);
